refactor(chat): extract message actions into hook
Move message + conversation action handlers from Chat.jsx into
src/hooks/useChatMessageActions.js: send (text/image/file), edit,
delete message, reply state wiring, and the confirmable chat actions
sharing the pendingChatAction modal (leave team, remove conversation,
delete message). execute* helpers stay internal to the hook; Chat.jsx
remains the state owner and passes state + setters in. Removes now-unused
socketService/teamService/uploadToImageKit imports. Chat.jsx 1350 -> 962.
